Time: 2 ms (62.94%), Space: 6.6 MB (68.30%) - LeetHub

+++ 377-combination-sum-iv/377-combination-sum-iv.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int combinationSum4(vector<int>& nums, int target) {
-        vector<long double>dp(target+1, 0);
+        vector<unsigned int>dp(target+1, 0);
         dp[0] = 1;
         for(int t = 1; t <= target; t++)
         {
